+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Table with data_getter Selects
 - Improved ag-grid cell selector
+- Allow passing additional styles to the model form container
 
 ## [0.8.0] - 2024-10-28
 ### Added